requirements.txt
gradio==3.42.0
matplotlib==3.7.2
pillow==9.5.0
groq
